// frontend/src/openapi/models/PipelineRun.ts
export interface PipelineRun {
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    pipeline: string | null;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    pipelineName: string | null;
    /**
     *
     * @type {number}
     * @memberof PipelineRun
     */
    revision: number | null;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    sourceDataset: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    sourceCell: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    sourceFingerprint: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    artifactCell: string | null;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    artifactFingerprint: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    provenance: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    execution: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    mode: string;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof PipelineRun
     */
    parameters: { [key: string]: any; };
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    binding: string | null;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof PipelineRun
     */
    operation: { [key: string]: any; } | null;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof PipelineRun
     */
    runner: { [key: string]: any; } | null;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    outputCell: string | null;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    state: string;
    /**
     *
     * @type {string}
     * @memberof PipelineRun
     */
    error: string;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof PipelineRun
     */
    result: { [key: string]: any; };
    /**
     *
     * @type {Date}
     * @memberof PipelineRun
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof PipelineRun
     */
    updatedAt: Date;
    /**
     *
     * @type {Date}
     * @memberof PipelineRun
     */
    completedAt: Date | null;
    /**
     *
     * @type {number}
     * @memberof PipelineRun
     */
    pollAfterSeconds: number | null;
    /**
     *
     * @type {number}
     * @memberof PipelineRun
     */
    queueSeconds: number | null;
}